requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.4
